Time: 0 ms (100%), Space: 13.6 MB (7.74%) - LeetHub

+++ 2692-take-gifts-from-the-richest-pile/2692-take-gifts-from-the-richest-pile.cpp
@@ -7,7 +7,7 @@ public:
             maxheap.pop();
             maxheap.push(floor(sqrt(n)));
         }
-        long int total = 0;
+        long long int total = 0;
         while (!maxheap.empty()) {
         total += maxheap.top();
         maxheap.pop();
